requirement.txt
Django==5.2.8
pdfplumber==0.11.8
sentence-transformers==5.1.2
chromadb==0.5.18
nltk==3.9.1
ollama==0.3.3
openai==1.65.2
python-dotenv==1.0.1
django-environ==0.12.0